feat: add base routing and reorganize App component
