+++ src/components/layout/Layout.tsx
@@ -1,9 +1,9 @@
-import type { LayoutProps } from '@/types/types';
 import Header from './Header';
 import Footer from './Footer';
 import { type JSX } from 'react';
+import { Outlet } from 'react-router';
 
-const Layout = ({ children }: LayoutProps): JSX.Element => {
+const Layout = (): JSX.Element => {
   return (
     <div className="px-5 flex flex-col items-center justify-center text-center mx-auto bg-gray-200 dark:bg-neutral-900 text-black dark:text-white min-h-full w-full">
       <Header />
@@ -11,7 +11,7 @@ const Layout = ({ children }: LayoutProps): JSX.Element => {
         role="main"
         className="container grow sm:gap-20 gap-10 flex flex-col justify-start items-center w-full"
       >
-        {children}
+        <Outlet />
       </main>
       <Footer />
     </div>
