+++ src/pages/MainPage.tsx
@@ -1,7 +1,23 @@
-import { type JSX } from 'react';
+import type { JSX } from 'react';
+import useFetchData from '@/hooks/useFetchData';
+import useLocalStorage from '@/hooks/useLocalStorage';
+import SearchResults from '@/components/results/SearchResults';
+import Search from '@/components/search/Search';
 
 const MainPage = (): JSX.Element => {
-  return <div>MainPage</div>;
+  const [searchQuery, setSearchQuery] = useLocalStorage('searchQuery', '');
+  const { items, isLoading, error } = useFetchData(searchQuery);
+
+  const handleSearch = (query: string): void => {
+    setSearchQuery(query);
+  };
+
+  return (
+    <>
+      <Search onSearch={handleSearch} searchQuery={searchQuery} />
+      <SearchResults items={items} error={error} isLoading={isLoading} searchQuery={searchQuery} />
+    </>
+  );
 };
 
 export default MainPage;
